Added Rotation component, normalized diagonal movement.
Changes:
Added a Rotation component to player entities that describes which
direction they're facing.

Added rotation updates to all places where we update movement.

Fixed a bug where the YDown/YUp inputs were flipped (YUp is southward).

Changed SharedConfig::MOVEMENT_VELOCITY from double to float since our
movement math and variables are all float.

Fixed an issue where diagonal movement was faster than cardinal movement.

+++ Source/SharedLib/Config/Public/SharedConfig.h
@@ -55,7 +55,7 @@ public:
 
     /** The speed that entities will move at. Eventually we'll support
         dynamically setting this based on stats, for now it's a constant. */
-    static constexpr double MOVEMENT_VELOCITY{50};
+    static constexpr float MOVEMENT_VELOCITY{50};
 
     //-------------------------------------------------------------------------
     // Network
